+++ app/include_upgrade_mainnet.go
@@ -11,7 +11,7 @@ import (
 // init is used to include v1 upgrade for mainnet data
 func init() {
 	Upgrades = []upgrades.Upgrade{v1.CreateUpgrade(v1.UpgradeDataString{
-		BtcStakingParamStr:        mainnet.BtcStakingParamStr,
+		BtcStakingParamsStr:       mainnet.BtcStakingParamsStr,
 		FinalityParamStr:          mainnet.FinalityParamStr,
 		IncentiveParamStr:         mainnet.IncentiveParamStr,
 		CosmWasmParamStr:          mainnet.CosmWasmParamStr,