+++ test/e2e/software_upgrade_e2e_v1_test.go
@@ -94,8 +94,8 @@ func (s *SoftwareUpgradeV1TestnetTestSuite) TearDownSuite() {
 	}
 }
 
-// TestUpgradeSignetLaunch Checks if the BTC Headers were inserted.
-func (s *SoftwareUpgradeV1TestnetTestSuite) TestUpgradeSignetLaunch() {
+// TestUpgradeV1 Checks if the BTC Headers were inserted.
+func (s *SoftwareUpgradeV1TestnetTestSuite) TestUpgradeV1() {
 	// chain is already upgraded, only checks for differences in state are expected
 	chainA := s.configurer.GetChainConfig(0)
 
@@ -145,10 +145,17 @@ func (s *SoftwareUpgradeV1TestnetTestSuite) TestUpgradeSignetLaunch() {
 	// as the one from the data
 	stakingParams := n.QueryBTCStakingParams()
 
-	stakingParamsFromData, err := v1.LoadBtcStakingParamsFromData(bbnApp.AppCodec(), testnet.BtcStakingParamStr)
+	bsParamsFromUpgrade, err := v1.LoadBtcStakingParamsFromData(testnet.BtcStakingParamsStr)
 	s.NoError(err)
 
-	s.EqualValues(stakingParamsFromData, *stakingParams)
+	lastParamInUpgradeData := bsParamsFromUpgrade[len(bsParamsFromUpgrade)-1]
+	s.EqualValues(lastParamInUpgradeData, *stakingParams)
+
+	// expected version starts at 0 since the upgrade overwrites the params
+	for expVersion, paramsInUpgradeData := range bsParamsFromUpgrade {
+		bsParamsAtBtcHeight := n.QueryBTCStakingParamsByVersion(uint32(expVersion))
+		s.Equal(*bsParamsAtBtcHeight, paramsInUpgradeData)
+	}
 
 	// check that finality params correctly deserialize and that they are the same
 	// as the one from the data