+++ x/incentive/types/keys.go
@@ -1,6 +1,10 @@
 package types
 
-import "cosmossdk.io/collections"
+import (
+	"cosmossdk.io/collections"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/address"
+)
 
 const (
 	// ModuleName defines the module name
@@ -19,10 +23,15 @@ const (
 var (
 	ParamsKey                            = []byte{0x01}             // key prefix for the parameters
 	BTCStakingGaugeKey                   = []byte{0x02}             // key prefix for BTC staking gauge at each height
-	ReservedKey                          = []byte{0x03}             // reserved //nolint:unused
+	DelegatorWithdrawAddrPrefix          = []byte{0x03}             // key for delegator withdraw address
 	RewardGaugeKey                       = []byte{0x04}             // key prefix for reward gauge for a given stakeholder in a given type
 	RefundableMsgKeySetPrefix            = collections.NewPrefix(5) // key prefix for refundable msg key set
 	FinalityProviderCurrentRewardsKey    = []byte{0x06}             // key prefix for storing the Current rewards of finality provider by addr
 	FinalityProviderHistoricalRewardsKey = []byte{0x07}             // key prefix for storing the Historical rewards of finality provider by addr and period
 	BTCDelegationRewardsTrackerKey       = []byte{0x8}              // key prefix for BTC delegation rewards tracker info (del,fp) => BTCDelegationRewardsTracker
 )
+
+// GetWithdrawAddrKey creates the key for a delegator's withdraw addr.
+func GetWithdrawAddrKey(delAddr sdk.AccAddress) []byte {
+	return append(DelegatorWithdrawAddrPrefix, address.MustLengthPrefix(delAddr.Bytes())...)
+}