+++ CHANGELOG.md
@@ -39,8 +39,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Improvements
 
+- [#309](https://github.com/babylonlabs-io/babylon/pull/309) feat(adr-036): custom withdrawal address
 - [#305](https://github.com/babylonlabs-io/babylon/pull/305) chore: add more error logs to `VerifyInclusionProofAndGetHeight`
 - [#304](https://github.com/babylonlabs-io/babylon/pull/304) Add highest voted height to finality provider
+- [#314](https://github.com/babylonlabs-io/babylon/pull/314) Require exact unbonding time in delegation
+
+### State Machine Breaking
+
+- [#310](https://github.com/babylonlabs-io/babylon/pull/310) implement adr-37 -
+making params valid for btc light client ranges
+
+## v0.17.2
+
+### Improvements
+- [#311](https://github.com/babylonlabs-io/babylon/pull/311) Enforce version 2
+for unbonding transactions
 
 ## v0.17.1